Implement the releases as a data source
Previously there was one huge file for all the versions in the
navigation. In addition to that, there were individual release pages.
This was difficult to maintain.

The new way uses Nanoc data sources[1] to essentially act as a database.
Based on that the various release pages are generated from a single
source. For that, guides are identified by tags so they can be renamed
without affecting the website. The releases on the index page are also
generated.

That allows easy copying when a release is branched and provides an easy
overview which builds belong to which release.

There is still some room for improvement. A lot of content in the
release file overlaps with the content in guides. If it could be
extracted from guides, that would reduce maintenance even further.

[1]: https://nanoc.app/doc/data-sources/